clean and restructure folders
