+++ index.html
@@ -9,9 +9,10 @@
 <link rel="icon" type="image/png" href="src/img/logo__mobile.png" />
 <link href="https://fonts.googleapis.com/css2?family=Roboto:wght@300;400;500&display=swap" rel="stylesheet">
 <link href="https://fonts.googleapis.com/css2?family=Comfortaa:wght@300;400;500;600;700&family=Roboto:wght@300;400;500&display=swap" rel="stylesheet">
-<script defer type="module" src="src/js/modelo.js"></script>
 <script defer type="module" src="src/js/show.js"></script>
+<script defer type="module" src="src/js/modelo.js"></script>
 <script defer type="module" src="src/js/seach.js"></script>
+<script defer type="module" src="src/js/gallery.js"></script>
 <body>
     <header class="container header_home">
         <div class="menu_mobile__logo">
@@ -50,16 +51,50 @@
                 </div>
             </nav>
         </div>
+        <div class="nav_new hidden">
+            <h3 class="h3__title">Create</h3>
+            <p >New Pic</p>
+            <p >New Table</p>
+        </div>
     </header>
     <div id="container_principal" class="container">
-        <!-- <form class="seach__wrapper" action="">
-            <div class="seach__logo">
-                <svg width="15" height="15" viewBox="0 0 15 15" fill="none" xmlns="http://www.w3.org/2000/svg">
-                    <path fill-rule="evenodd" clip-rule="evenodd" d="M12.7867 6.5C12.7867 9.53757 10.3242 12 7.28668 12C4.24912 12 1.78668 9.53757 1.78668 6.5C1.78668 3.46243 4.24912 1 7.28668 1C10.3242 1 12.7867 3.46243 12.7867 6.5ZM11.5158 11.4362C10.379 12.411 8.90163 13 7.28668 13C3.69683 13 0.786682 10.0899 0.786682 6.5C0.786682 2.91015 3.69683 0 7.28668 0C10.8765 0 13.7867 2.91015 13.7867 6.5C13.7867 8.11495 13.1977 9.59234 12.2229 10.7291L16.6402 15.1464L15.9331 15.8536L11.5158 11.4362Z" fill="#645D59"/>
-                    </svg>
+        <!-- <section id="user hidden">
+            <div class="profile__info">
+                <div class="info__img">
+                    <img src="src/img/foto3.jpg" alt="">
+                </div>
+                <h2 class="h2__title">Verónica Velázquez</h2>
+                <p class="pin__user">@verovelazquez</p>
             </div>
-            <input class="seach__input" type="text" placeholder="Seach">
-        </form> -->
+            <div class="container profile__table">
+                <div class="table__title">
+                    <p>ALL TABLES</p>
+                    <p>ALL PINS</p>
+                </div>
+                <div class="table__tables">
+                    <div class="tables__table"></div>
+                    <div class="tables__table"></div>
+                    <div class="tables__table"></div>
+                    <div class="tables__table"></div>
+                </div>
+            </div>
+        </section> -->
+        <!-- <section id="form-login" class="form__registration">
+            <form action="" clas>
+                <div class="modal-button">
+                    <button id="goBack" class="button__goback"><svg width="25px" height="25px" viewBox="0 0 25 25" fill="none" xmlns="http://www.w3.org/2000/svg">
+                        <path fill-rule="evenodd" clip-rule="evenodd" d="M1.91421 4.35355L4.56066 7L3.85355 7.70711L0.353553 4.20711L0 3.85355L0.353553 3.5L3.85355 0L4.56066 0.707107L1.91421 3.35355L5.70711 3.35355C9.28325 3.35355 12.2071 6.27741 12.2071 9.85355L12.2071 11.3536H11.2071L11.2071 9.85355C11.2071 6.8297 8.73096 4.35355 5.70711 4.35355H1.91421Z" fill="#5B5551"/>
+                        </svg></button>
+                </div>
+                <div class="container">
+                    <h2 class="h1_title">Register</h2>
+                    <input class="input__login" type="text" placeholder="jane@example.com">
+                    <input class="input__login" type="text" placeholder="*******">
+                    <input class="button__send" type="submit" value="NEXT">
+                </div>
+
+            </form>
+        </section> -->
         <section id="gallery" class="grid-gallery">
         </section>
     </div>
