+++ src/css/buttons.css
@@ -8,3 +8,12 @@
     background-color: transparent;
     flex-flow: column wrap;
 }
+.button__send {
+    width: 100%;
+    height: 40px;
+    border-radius: 20px;
+    color: var(--color-secundary);
+    background-color: var(--background-principal);
+    border: 1px solid var(--color-secundary);
+    padding: 10px;
+}
