+++ src/css/form.css
@@ -1,0 +1,13 @@
+.form__registration {
+    width: 100%;
+    height: 80vh;
+}
+.input__login {
+    width: 100%;
+    height: 40px;
+    border-radius: 20px;
+    margin-bottom: 20px;
+    color: var(--color-secundary);
+    border: 1px solid var(--color-secundary);
+    padding: 10px;
+}
