+++ src/css/modal.css
@@ -29,7 +29,7 @@
     background-size: cover;
     padding: 10px 5px 0 5px;
 }
-.modal-pin .modal-pin__user{
+.modal-pin .pin__user{
     width: 100%;
     height: 20px;
     margin: 10px;
