+++ src/js/model/user.js
@@ -1,0 +1,15 @@
+export const users = [
+    {
+        id: 0,
+        user: '@verovelazquez',
+        name: 'Verónica Velázquez',
+        photoprofile: 'src/img/building/imagen1.jpg',
+        tables: [
+            {
+                tname: 'Tile Table',
+                pins: ['src/img/building/imagen1.jpg', 'src/img/building/imagen1.jpg', 'src/img/building/imagen1.jpg', 'src/img/building/imagen1.jpg']
+            }
+        ],
+        pins: ['src/img/building/imagen1.jpg', 'src/img/building/imagen1.jpg', 'src/img/building/imagen1.jpg', 'src/img/building/imagen1.jpg']
+    }
+]
